requirements.txt
fastapi
uvicorn

streamlit

torch
transformers
sentencepiece

chromadb
sentence-transformers

nltk
tiktoken
pandas
pyyaml
requests

python-jose
passlib[bcrypt]
bcrypt==4.0.1

python-multipart
